feat/test: 피드백 반영하여 수정
* save() 시 id가 return 되도록 수정
* 테스트 간 데이터베이스 격리를 위해 reset-db.sql 작성 및 적용

+++ src/main/java/nextstep/courses/domain/repository/SessionRepository.java
@@ -5,7 +5,7 @@ import nextstep.courses.domain.Session;
 import java.util.Optional;
 
 public interface SessionRepository {
-    int save(Session session);
+    Long save(Session session);
 
     Optional<Session> findById(Long id);
 }
