+++ src/main/java/nextstep/courses/domain/repository/StudentRepository.java
@@ -5,7 +5,7 @@ import nextstep.courses.domain.Student;
 import java.util.List;
 
 public interface StudentRepository {
-    int save(Long userId, Long sessionId);
+    Long save(Long userId, Long sessionId);
 
     List<Student> findAllBySessionId(Long sessionId);
 }
